Split sign-bundle logic into 2 with signType flag
