+++ go/bundle/cmd/sign-bundle/integrityblock.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"crypto"
+)
+
+func SignIntegrityBlock(privKey crypto.PrivateKey) error {
+	// TODO(sonkkeli): Actual logic.
+
+	return nil
+}
